pyproject.toml
[tool.poetry]
name = "analytical-pipeline"
version = "0.0.1"
description = "real-time stocks market analytical pipeline"
authors = ["yuto82 <rgritsal@gmail.com>"]
readme = "README.md"
packages = [{include = "analytical_pipeline", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.32.0"
websocket-client = "^1.8.0"
pandas = "^2.3.0"
numpy = "^2.3.0"
colorlog = "^6.9.0"
kafka-python = "^2.2.15"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]